Fix Int Casting Issue
- Issue preventing view results to show up after faculty clicks Check Plagiarism.

+++ backend/src/main/java/com/womm/backend/service/PlagiarismService.java
@@ -43,7 +43,7 @@ public class PlagiarismService {
         }
 
         results.sort((x, y) ->
-                Integer.compare((int) y.get("similarity"), (int) x.get("similarity"))
+                Long.compare((Long) y.get("similarity"), (Long) x.get("similarity"))
         );
 
         return results;
